requirements.txt
flask==3.0.2
groq>=0.4.2
python-dotenv==1.0.1
requests==2.31.0
pyttsx3>=2.90
SpeechRecognition>=3.10.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
streamlit>=1.28.0
pyaudio>=0.2.13 